// my-project/src/client/schema.ts
export const backgroundImageSchema = mediaAssetSchema.extend({});
export type BackgroundImage = z.infer<typeof backgroundImageSchema>;

export const cgImageSchema = mediaAssetSchema.extend({});
export type CGImage = z.infer<typeof cgImageSchema>;

export const soundEffectSchema = mediaAssetSchema.extend({});
export type SoundEffect = z.infer<typeof soundEffectSchema>;

export const bgmSchema = mediaAssetSchema.extend({});
export type BGM = z.infer<typeof bgmSchema>;

/* ------------------------------------------------------
    Character
------------------------------------------------------ */
export const characterSchema = z.object({
	id: z.string(),
	name: z.string(),
	images: z.record(z.string(), characterImageSchema),
});
export type Character = z.infer<typeof characterSchema>;

/* ------------------------------------------------------
    GameResources
------------------------------------------------------ */
export const gameResourcesSchema = z.object({
	characters: z.record(z.string(), characterSchema),
	backgroundImages: z.record(z.string(), backgroundImageSchema),
	soundEffects: z.record(z.string(), soundEffectSchema),
	bgms: z.record(z.string(), bgmSchema),
	cgImages: z.record(z.string(), cgImageSchema),
});
export type GameResources = z.infer<typeof gameResourcesSchema>;